chore(docs): move to /packages
