+++ packages/administration/src/docs/docs.component.ts
