+++ packages/administration/src/docs/docs.module.ts
