+++ packages/administration/src/docs/index.ts
@@ -1,0 +1,2 @@
+export { DocsModule } from './docs.module';
+export { JhiDocsComponent } from './docs.component';
