--- src/app/admin/docs/index.ts
@@ -1,1 +1,0 @@
-export { DocsModule } from './docs.module';
